move NettyUtils to tachyon.worker.netty; may move it to common utils when netty is more widely used.
