+++ core/src/main/java/tachyon/worker/netty/NettyDataServer.java
@@ -29,7 +29,6 @@ import com.google.common.base.Throwables;
 
 import tachyon.Constants;
 import tachyon.conf.TachyonConf;
-import tachyon.util.NettyUtils;
 import tachyon.worker.BlocksLocker;
 import tachyon.worker.DataServer;
